#40. Fix some of the -Weffc++ warnings.
git-svn-id: https://subversion.gsi.de/fairroot/cbmroot/trunk@9900 0381ead4-6506-0410-b988-94b70fbc4730

+++ littrack/base/CbmLitFloat.h
@@ -11,6 +11,6 @@
 #ifndef CBMLITFLOAT_H_
 #define CBMLITFLOAT_H_
 
-typedef double myf;
+typedef float myf;
 
 #endif /*CBMLITFLOAT_H_*/
